functions/kintone-pusher/requirements.txt
# ==== Google Cloud SDK ====
google-cloud-storage==2.14.0
functions-framework==3.5.0

# ==== HTTP & Utilities ====
requests==2.31.0
python-dotenv==1.0.0